+++ src/app/modules/step-3-view-cohort-page/view-cohort-page.component.spec.ts
@@ -28,7 +28,6 @@ describe('ViewCohortComponent', () => {
         {
           provide: ColumnDescriptionsService,
           useValue: {
-            getWrapCellText: () => false,
             getVisibleColumns: () => of([]),
             destroy: () => {}
           }