+++ src/app/modules/stepper/stepper.component.spec.ts
@@ -90,7 +90,6 @@ describe('StepperComponent', () => {
         {
           provide: ColumnDescriptionsService,
           useValue: {
-            getWrapCellText: () => false,
             getVisibleColumns: () => of([])
           }
         }